notes-app-backend/requirements.txt
Django
psycopg2-binary
djangorestframework
django-cors-headers
python-jose[cryptography]
requests
